keep best merge and best.log

+++ sd_webui_bayesian_merger/merger.py
@@ -62,10 +62,14 @@ class Merger:
 
     def remove_previous_ckpt(self, current_it: int) -> None:
         if current_it > 1:
-            self.create_model_out_name(current_it - 1)
-            print(f"Removing {self.output_file}")
-            self.output_file.unlink()
+            if self.output_file.exists():
+                self.create_model_out_name(current_it - 1)
+                print(f"Removing {self.output_file}")
+                self.output_file.unlink()
         self.create_model_out_name(current_it)
+
+    def keep_best_ckpt(self) -> None:
+        self.output_file.rename(self.best_output_file)
 
     def merge(
         self,
